some changes here and there

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: eburnet <eburnet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 11:06:52 by eburnet           #+#    #+#             */
-/*   Updated: 2023/11/07 13:00:28 by eburnet          ###   ########.fr       */
+/*   Updated: 2023/11/08 19:44:38 by eburnet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ char	*ft_strchr(const char *s, int c);
 char	*ft_strdup(const char *s);
 void	ft_striteri(char *s, void (*f)(unsigned int, char *));
 char	*ft_strjoin(char const *s1, char const *s2);
-int		ft_strlcat(char *dest, const char *src, int n);
+size_t	ft_strlcat(char *dest, const char *src, size_t n);
 int		ft_strlcpy(char *dst, const char *src, size_t size);
 int		ft_strlen(char *str);
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
